requirements.txt
2to3==1.0
altgraph==0.16.1
asn1crypto==0.24.0
autoflake==1.2
beautifulsoup4==4.7.1
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
chromedriver-binary==2.45.0
Click==7.0
colorama==0.4.1
comtypes==1.1.7
cryptography==2.4.2
Cython==0.29.2
et-xmlfile==1.0.1
future==0.17.1
idna==2.8
jdcal==1.4
lxml==4.3.0
macholib==1.11
numpy==1.15.4
opencv-python==4.0.0.21
openpyxl==2.5.12
packaging==18.0
pefile==2018.8.8
Pillow==5.4.1
pip-review==1.0
PyAutoGUI==0.9.41
pycparser==2.19
pyflakes==2.0.0
pygame==1.9.4
PyGetWindow==0.0.3
PyInstaller==3.4
PyMsgBox==1.0.6
pyOpenSSL==18.0.0
pyparsing==2.3.0
pypng==0.0.19
PyQRCode==1.2.1
PyQt5==5.11.3
PyQt5-sip==4.19.13
pyqt5-tools==5.11.3.1.4
PyRect==0.1.4
PyScreeze==0.1.19
PySocks==1.6.8
pytesseract==0.2.6
python-dateutil==2.7.5
python-dotenv==0.10.1
PyTweening==1.0.3
pytz==2018.9
pywin32==224
pywin32-ctypes==0.2.0
pywinauto==0.6.5
pyzbar==0.1.7
qrcode==6.0
reedsolo==0.3
requests==2.21.0
selenium==3.141.0
six==1.12.0
soupsieve==1.7
SpeechRecognition==3.8.1
urllib3==1.24.1
virtualenv==16.2.0
wget==3.2
wincertstore==0.2
XlsxWriter==1.1.2
xlwt==1.3.0
